Update FLOOD_MAPPING_SCRIPT_V1.js
I have updated the code with an improved version of the same. It is well commented. As for the ROI, I shall share it with you right away. Find the link to the code as at this time below:
https://code.earthengine.google.com/e7bbc3a6ef8fa29607f4744bd05542c1
This code only marks the first step, which is simply obtaining the SAR imagery, filtering and processing the imagery, accordingly, obtaining threshold values and masking to obtain flooded regions. I shall update you guys on the next phase which is integrating NDWI from Sentinel-2 MSI and share the corresponding link to the same accordingly. Rest assured